# example_main/docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  response.RespInfo:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  license:
    name: XXX copyright
  title: XXX Service APIs
  version: "1.0"
paths:
  /example/get/{id}:
    get:
      consumes:
      - application/json
      description: get Example Object by ID
      parameters:
      - description: Example ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RespInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RespInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RespInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RespInfo'
      summary: get A Example
      tags:
      - example
  /example/on-async-task/get/{id}:
    get:
      consumes:
      - application/json
      description: get Example Object by ID, on Dispatcher a task to do something
      parameters:
      - description: Example ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RespInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RespInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RespInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RespInfo'
      summary: get A Example with Dispatcher Task
      tags:
      - example
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"